flask_backend/requirements.txt
Flask==3.1.1
flask_cors==5.0.1
pytest==8.3.5
python-dotenv==1.1.0
Requests==2.32.3
